Allow use of document.body
